+++ fiboa_cli/convert_utils.py
@@ -16,6 +16,8 @@ import sys
 import zipfile
 import py7zr
 import flatdict
+import rarfile
+
 
 def convert(
         output_file, cache_path,
@@ -38,7 +40,7 @@ def convert(
         license = None,
         compression = None,
         geoparquet1 = False,
-        explode_multipolygon = False,
+        original_geometries = False,
         index_as_id = False,
         **kwargs):
     """
@@ -158,11 +160,12 @@ def convert(
             else:
                 log(f"Column '{key}' not found in dataset, skipping migration", "warning")
 
-    # 4b. For geometry column, convert multipolygon type to polygon
-    if explode_multipolygon:
+    # 4b. For geometry column, fix geometries
+    if not original_geometries:
+        gdf.geometry = gdf.geometry.make_valid()
         gdf = gdf.explode()
 
-    if has_migration or has_col_migrations or has_col_filters or has_col_additions or explode_multipolygon:
+    if has_migration or has_col_migrations or has_col_filters or has_col_additions:
         log("GeoDataFrame after migrations and filters:")
         print(gdf.head())
 
@@ -403,7 +406,6 @@ def download_files(uris, cache_folder = None):
                 with py7zr.SevenZipFile(cache_file, 'r') as sz_file:
                     sz_file.extractall(zip_folder)
             elif name.endswith(".rar"):
-                import rarfile
                 with rarfile.RarFile(cache_file, 'r') as w:
                     w.extractall(zip_folder)
             else: