+++ fiboa_cli/parquet.py
@@ -5,13 +5,20 @@ from geopandas import GeoDataFrame
 from shapely.geometry import shape
 
 from .types import get_geopandas_dtype, get_pyarrow_type_for_geopandas, get_pyarrow_field
-from .util import log, load_fiboa_schema, load_file, merge_schemas
+from .util import log, load_fiboa_schema, load_file, merge_schemas, is_schema_empty
 from .geopandas import to_parquet
+
+ROW_GROUP_SIZE = 25000
 
 def create_parquet(data, columns, collection, output_file, config, missing_schemas = {}, compression = None, geoparquet1 = False):
     # Load the data schema
     fiboa_schema = load_fiboa_schema(config)
     schemas = merge_schemas(missing_schemas, fiboa_schema)
+
+    # Add the custom schemas to the collection for future use
+    if not is_schema_empty(missing_schemas):
+        collection = collection.copy()
+        collection["fiboa_custom_schemas"] = missing_schemas
 
     # Load all extension schemas
     extensions = {}
@@ -84,6 +91,7 @@ def create_parquet(data, columns, collection, output_file, config, missing_schem
         coerce_timestamps = "ms",
         compression = compression,
         schema_version = "1.0.0" if geoparquet1 else "1.1.0",
+        row_group_size = ROW_GROUP_SIZE,
         write_covering_bbox = False if geoparquet1 else True
     )