+++ CHANGELOG.md
@@ -9,6 +9,38 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Command `fiboa improve` with helpers to
+  - change the CRS
+  - change the GeoParquet version and compression
+  - fill missing perimeter/area values
+  - fix invalid geometries
+  - rename columns
+- Converter for Lithuania (EuroCrops)
+- Converter for Slovenia
+- Converter for Slovakia
+- Converter for Switzerland
+- `fiboa convert`: New parameter `--original-geometries` / `-og` to keep the original geometries
+
+### Changed
+
+- `fiboa convert`:
+  - Writes custom schemas to collection metadata
+  - Geometries are made valid using GeoPanda's `make_valid` method by default
+  - MultiPolygons are converted to Polygons by default
+- `fiboa validate` uses custom schemas for validation
+- `fiboa merge` keeps custom schemas when needed
+
+### Removed
+- `fiboa convert`: Removed the explicit parameter `explode_multipolygon` from the converter
+
+### Fixed
+
+- Fix converter for Estland to use the id `ec_ee` instead of `ec_es`
+
+## [v0.8.0] - 2024-11-12
+
+### Added
+
 - Merge command: `fiboa merge`
 - Converter for Croatia
 - Converter for Germany, Mecklenburg-Western Pomerania
@@ -18,10 +50,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Converter for Sweden
 - Converter for Luxembourg
 - Converter for Ireland
+- Converter for Lacuna Labels (A region-wide, multi-year set of crop field boundary labels for Africa)
 
 ### Changed
 
 - `fiboa convert`: Default compression changed from `zstd` to `brotli`
+- The default row group size of exported parquet files was changed from ~1.000.000 to 25.000
 
 ### Fixed
 
@@ -327,7 +361,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - First release
 
-[Unreleased]: <https://github.com/fiboa/cli/compare/v0.7.0...main>
+[Unreleased]: <https://github.com/fiboa/cli/compare/v0.8.0...main>
+[v0.8.0]: <https://github.com/fiboa/cli/compare/v0.7.0...v0.8.0>
 [v0.7.0]: <https://github.com/fiboa/cli/compare/v0.6.0...v0.7.0>
 [v0.6.0]: <https://github.com/fiboa/cli/compare/v0.5.0...v0.6.0>
 [v0.5.0]: <https://github.com/fiboa/cli/compare/v0.4.0...v0.5.0>